+++ test/extensions/transport_sockets/alts/alts_integration_test.cc
@@ -204,7 +204,7 @@ INSTANTIATE_TEST_SUITE_P(IpVersions, AltsIntegrationTestClientInvalidPeer,
 // any account in config, the handshake will fail and client closes connection.
 TEST_P(AltsIntegrationTestClientInvalidPeer, clientValidationFail) {
   initialize();
-  codec_client_ = makeRawHttpConnection(makeAltsConnection());
+  codec_client_ = makeRawHttpConnection(makeAltsConnection(), absl::nullopt);
   EXPECT_FALSE(codec_client_->connected());
 }
 
@@ -252,7 +252,7 @@ INSTANTIATE_TEST_SUITE_P(IpVersions, AltsIntegrationTestClientWrongHandshaker,
 // and connection closes.
 TEST_P(AltsIntegrationTestClientWrongHandshaker, ConnectToWrongHandshakerAddress) {
   initialize();
-  codec_client_ = makeRawHttpConnection(makeAltsConnection());
+  codec_client_ = makeRawHttpConnection(makeAltsConnection(), absl::nullopt);
   EXPECT_FALSE(codec_client_->connected());
 }